backend/requirements.txt
# Django Core
Django==6.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Database
psycopg[binary]==3.3.0

# Environment Variables
python-dotenv==1.0.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-magic-bin==0.4.14  # For Windows, use python-magic on Linux

# Security
django-encrypted-model-fields==0.6.5
cryptography==41.0.7

# API Rate Limiting
django-ratelimit==4.1.0

# Audit Logging
django-auditlog==2.3.0

# PDF Generation
reportlab==4.0.8
xhtml2pdf==0.2.11

# LLM Integration
openai==1.6.1
anthropic==0.8.1
tiktoken==0.5.2

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
Pillow==12.1.0
celery==5.3.4
redis==5.0.1
